+++ src/main/java/com/example/waytogo/routes_points/service/impl/RoutePointServiceJPA.java
@@ -3,14 +3,17 @@ package com.example.waytogo.routes_points.service.impl;
 import com.example.waytogo.routes_points.entity.RoutePoint;
 import com.example.waytogo.routes_points.repository.RoutePointRepository;
 import com.example.waytogo.routes_points.service.api.RoutePointService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@Validated
 @RequiredArgsConstructor
 public class RoutePointServiceJPA implements RoutePointService {
     private final RoutePointRepository routePointRepository;
@@ -27,7 +30,7 @@ public class RoutePointServiceJPA implements RoutePointService {
     }
 
     @Override
-    public RoutePoint saveNewRoutePoint(RoutePoint routePoint) {
+    public RoutePoint saveNewRoutePoint(@Valid RoutePoint routePoint) {
 
         return routePointRepository.save(routePoint);
     }
@@ -40,8 +43,13 @@ public class RoutePointServiceJPA implements RoutePointService {
     }
 
     @Override
-    public void deleteRoutePointById(UUID routePointId) {
-        routePointRepository.deleteById(routePointId);
+    public Boolean deleteRoutePointById(UUID routePointId) {
+        if(routePointRepository.existsById(routePointId))
+        {
+            routePointRepository.deleteById(routePointId);
+            return true;
+        }
 
+        return false;
     }
 }
